requirements.txt
requests
python-dotenv
SQLAlchemy>=2.0
PyMySQL
pymongo
